requirements.txt
openai
python-dotenv
mdutils